replace defcusom `request-coding-system` with key `:coding-system`
Not tested yet!

This code removes the new `request-coding-system` defcustom, and
instead uses a keyword parameter to the main `request` function. Since
this parameter needs to be passed down the chain several times I'm not
entirely certain that I've made the changes everywhere they need to be
made.

This commit also shifts the default form 'binary to 'utf-8.