Fix errors with using curl for file: URLs
Do not try to parse headers if the backend is curl and the URL does not
have the "http" or "https" scheme, and tolerate a nil value for the
response status code.  Curl does not print headers for "file:" or other
non-HTTP URLs.

* request.el (request--parse-data): Use `equal', which tolerates nil, to
compare the response status code.
(request--callback): Do not try to parse headers if using curl and not
using "http:" or "https:".
(request--curl-callback): Check whether the response status code is nil.