Add fix for an edge case (with a test) involving "100 Continue" and long body
Without that fix, Emacs interprets curl output as utf-8-dos and strips all \r .
It results in an error "(search-failed "^\r\n"),
when moving from "100 Continue" to the next one in request--goto-next-body.

It's kind of strange behaviour of Emacs,
since there is a single newline after the body:
body\n
(:num-redirects 0 :url-effective "http://127.0.0.1:60079/longtextline")
From the docs:

Emacs recognizes which kind of end-of-line conversion to use based on the
contents of the file: if it sees only carriage returns, or only carriage return
followed by linefeed sequences, then it chooses the end-of-line conversion accordingly.

This behaviour only manifests itself when body is a single long line.