Shorten accept-process-output interval
The curl process may not produce output, so we incur the full cost of the
accept-process-output timeout (formerly 300 ms).

We can't just eliminate the accept-process-output call since we need to yield control back
to the main thread which needs to reap the curl process.

We ought not to eliminate the timeout, as that frequently leads to CPU pegging.

So shorten the interval to 50 ms.  Empirically, querying over consumer
internet in the US takes 100 ms.  Such is the price of using `:sync t`.